requirements.txt
Django==4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2
django-split-settings==1.3.1